+++ db/migrate/20190623205308_remove_slug_from_reports.rb
@@ -1,0 +1,6 @@
+class RemoveSlugFromReports < ActiveRecord::Migration[6.0]
+  def change
+    remove_index :reports, :slug
+    remove_column :reports, :slug, :string
+  end
+end